requirements.txt
django
djangorestframework
django-cors-headers
django-rest-framework-nested
djangorestframework-simplejwt
drf-nested-routers
gunicorn
whitenoise
pillow
pika
drf_yasg
psycopg2-binary